#22 - BoardDB 테스트 관련 VO 및 테이블 수정
`BoardDBTest` 테스트 진행 중 쿼리 수행 불가 문제확인 되어 관련된 VO 및 테이블 일부 수정

+++ servlet-project-member-management/src/main/java/com/member/servletprojectmembermanagement/vo/BoardVo.java
@@ -24,7 +24,7 @@ public class BoardVo {
     private String ip;
     private Integer rippleCnt;  //게시판에 등록된 댓글의 갯수
     private String fileName;    //파일명
-    private Integer fileSize;    //파일크기
+    private Long fileSize;    //파일크기
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
